Small refactoring -  Move UnsignedLongParameter to datatypes (#8001)
