+++ datatypes/src/main/java/org/hyperledger/besu/datatypes/parameters/UInt256Parameter.java
@@ -12,20 +12,31 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.ethereum.api.jsonrpc.internal.parameters;
+package org.hyperledger.besu.datatypes.parameters;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import org.apache.tuweni.units.bigints.UInt256;
 
+/** A parameter that represents a UInt256 value. */
 public class UInt256Parameter {
 
   private final UInt256 value;
 
+  /**
+   * Create a new UInt256Parameter
+   *
+   * @param value the value
+   */
   @JsonCreator
   public UInt256Parameter(final String value) {
     this.value = UInt256.fromHexString(value);
   }
 
+  /**
+   * Get the value
+   *
+   * @return the value
+   */
   public UInt256 getValue() {
     return value;
   }
